+++ client/src/features/CasperCustom/CasperCustomPanel/CustomOptionImageItem.tsx
