+++ client/src/features/CasperCustom/CasperCustomPanel/EyesOptionImageItem.tsx
