+++ client/src/features/CasperCustom/CustomProcess/CasperCustomFinishing.tsx
@@ -2,11 +2,11 @@ import { useEffect, useState } from "react";
 import { motion } from "framer-motion";
 import { CASPER_SIZE_OPTION } from "@/constants/CasperCustom/casper";
 import { DISSOLVE } from "@/constants/animation";
+import { SCROLL_MOTION } from "@/constants/animation";
+import { CasperFlipCard } from "@/features/CasperCustom/CasperCard/CasperFlipCard";
+import { CasperCardType } from "@/features/CasperShowCase/TransitionCasperCards";
 import useCasperCustomStateContext from "@/hooks/useCasperCustomStateContext";
 import useToast from "@/hooks/useToast";
-import { SCROLL_MOTION } from "../../constants/animation";
-import { CasperCardType } from "../CasperShowCase/TransitionCasperCards";
-import { CasperFlipCard } from "./CasperFlipCard";
 
 interface CasperCustomFinishingProps {
     navigateNextStep: () => void;
