+++ client/src/features/CasperCustom/CustomProcess/CasperCustomProcess.tsx
@@ -4,10 +4,10 @@ import CTAButton from "@/components/CTAButton";
 import ListStep from "@/components/ListStep";
 import { CUSTOM_OPTION_ARRAY } from "@/constants/CasperCustom/customStep";
 import { DISSOLVE } from "@/constants/animation";
-import { MyCasperCardFront } from "@/features/CasperCustom/MyCasperCardFront";
+import { SCROLL_MOTION } from "@/constants/animation";
+import { MyCasperCardFront } from "@/features/CasperCustom/CasperCard/MyCasperCardFront";
 import useCasperCustomStateContext from "@/hooks/useCasperCustomStateContext";
 import { getCasperOptionDescription } from "@/utils/CasperCustom/getCasperOptionDescription";
-import { SCROLL_MOTION } from "../../constants/animation";
 
 interface CasperCustomProcessProps {
     handleClickNextStep: () => void;
