+++ client/src/features/CasperShowCase/TransitionCasperCards.tsx
@@ -2,9 +2,9 @@ import { useEffect, useRef, useState } from "react";
 import { AnimatePresence, motion, useAnimation } from "framer-motion";
 import { CASPER_CARD_SIZE, CASPER_SIZE_OPTION } from "@/constants/CasperCustom/casper";
 import { CARD_TRANSITION } from "@/constants/CasperShowCase/showCase";
+import { CasperFlipCard } from "@/features/CasperCustom/CasperCard/CasperFlipCard";
 import useLazyLoading from "@/hooks/useLazyLoading";
 import { SelectedCasperIdxType } from "@/types/casperCustom";
-import { CasperFlipCard } from "../CasperCustom/CasperFlipCard";
 
 export interface CasperCardType {
     id: number;
